Move OSInfo out of types.ts.
